authentification/requirements.txt
SQLAlchemy~=2.0.8
uvicorn~=0.21.1
fastapi~=0.95.0
pydantic~=1.10.7
python-dotenv~=1.0.0
PyJWT~=2.6.0
passlib~=1.7.4
bcrypt~=4.0.1